Removed a lot of the old functionality slowly phasing in new code
