--- Assets/Scripts/UI/InGameUI/PerformanceHUD/BasePerformanceComponent.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 3061dc0ae8a81814b81175e421a1e4c3
-timeCreated: 1476775597
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
