--- Assets/Scripts/UI/InGameUI/PerformanceHUD/BasePerformanceComponentUI.cs
@@ -1,18 +1,0 @@
-﻿#region License
-// ====================================================
-// Project Porcupine Copyright(C) 2016 Team Porcupine
-// This program comes with ABSOLUTELY NO WARRANTY; This is free software, 
-// and you are welcome to redistribute it under certain conditions; See 
-// file LICENSE, which is part of this source code package, for details.
-// ====================================================
-#endregion
-using UnityEngine;
-
-// Just so we can implement other UIs in the future (other than just TextPerformanceComponentUI)
-
-/// <summary>
-/// Just a base class for UI elements to subclass from.
-/// </summary>
-public abstract class BasePerformanceComponentUI : MonoBehaviour
-{
-}
