+++ Assets/Scripts/UI/InGameUI/UIElements.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 337f103be36ea164b941d6578e85007d
+folderAsset: yes
+timeCreated: 1485587755
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
