+++ Assets/Scripts/UI/InGameUI/UIElements/BaseUIElement.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 2df58ac3458e43141bbe65fe835c07e4
-timeCreated: 1476778388
+guid: 1394f0df30d193445a61854ccbe4f199
+timeCreated: 1485587756
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
